Move spark fixture into spark subpackage.
